Move type pareser to a separate file
